Move component to their own files, remove use of className in Item component
