+++ src/components/resources/Claim/Claim.js
@@ -27,6 +27,10 @@ import {
   Chevron,
 } from '../../ui';
 import Accordion from '../../containers/Accordion';
+import CareTeam from './CareTeam';
+import Diagnosis from './Diagnosis';
+import Insurance from './Insurance';
+import Items from './Items';
 
 const commonDTO = fhirResource => {
   const id = _get(fhirResource, 'id');
@@ -299,256 +303,6 @@ const resourceDTO = (fhirVersion, fhirResource) => {
   }
 };
 
-const CareTeam = props => {
-  const { careTeam } = props;
-
-  return (
-    <ValueSection label="Care Team" data-testid="careTeam" marginTop>
-      <Table>
-        <thead>
-          <TableRow>
-            <TableHeader>Provider</TableHeader>
-            <TableHeader>Role</TableHeader>
-            <TableHeader>Qualification</TableHeader>
-          </TableRow>
-        </thead>
-        <tbody className="border-top-0">
-          {careTeam.map((member, idx) => (
-            <TableRow key={idx}>
-              <TableCell data-testid="careTeam.provider" className="col-md-4">
-                <Reference fhirData={member.provider} />
-              </TableCell>
-              <TableCell data-testid="careTeam.role" className="col-md-4">
-                {member.role ? (
-                  <Coding fhirData={member.role} />
-                ) : (
-                  <MissingValue />
-                )}
-              </TableCell>
-              <TableCell
-                data-testid="careTeam.qualification"
-                className="col-md-4"
-              >
-                {careTeam.qualification ? (
-                  <Coding fhirData={member.qualification} />
-                ) : (
-                  <MissingValue />
-                )}
-              </TableCell>
-            </TableRow>
-          ))}
-        </tbody>
-      </Table>
-    </ValueSection>
-  );
-};
-
-const Diagnosis = props => {
-  const { diagnosis } = props;
-
-  return (
-    <ValueSection label="Diagnosis" data-testid="diagnosis" marginTop>
-      <Table>
-        <thead>
-          <TableRow>
-            <TableHeader>Diagnosis</TableHeader>
-            <TableHeader>Type</TableHeader>
-            <TableHeader>Package code</TableHeader>
-          </TableRow>
-        </thead>
-        <tbody className="border-top-0">
-          {diagnosis.map((diagnosis, idx) => (
-            <TableRow key={idx}>
-              <TableCell data-testid="diagnosis.diagnosis" className="col-md-4">
-                {diagnosis.coding ? (
-                  <Coding fhirData={diagnosis.coding} />
-                ) : diagnosis.refrence ? (
-                  <Reference fhirData={diagnosis.reference} />
-                ) : (
-                  <MissingValue />
-                )}
-              </TableCell>
-              <TableCell data-testid="diagnosis.type" className="col-md-4">
-                {diagnosis.typeCoding ? (
-                  <Coding fhirData={diagnosis.typeCoding} />
-                ) : (
-                  <MissingValue />
-                )}
-              </TableCell>
-              <TableCell
-                data-testid="diagnosis.packageCode"
-                className="col-md-4"
-              >
-                {diagnosis.packageCodeCoding ? (
-                  <Coding fhirData={diagnosis.packageCodeCoding} />
-                ) : (
-                  <MissingValue />
-                )}
-              </TableCell>
-            </TableRow>
-          ))}
-        </tbody>
-      </Table>
-    </ValueSection>
-  );
-};
-
-const Insurance = props => {
-  const { insurance } = props;
-
-  return (
-    <ValueSection label="Insurance" data-testid="insurance" marginTop>
-      <Table>
-        <thead>
-          <TableRow>
-            <TableHeader>Coverage</TableHeader>
-            <TableHeader>Business Arrangement</TableHeader>
-            <TableHeader>Claim Response</TableHeader>
-          </TableRow>
-        </thead>
-        <tbody className="border-top-0">
-          {insurance.map((insurance, idx) => (
-            <TableRow key={idx}>
-              <TableCell data-testid="insurance.coverage" className="col-md-4">
-                <Reference fhirData={insurance.coverage} />
-                {insurance.focal && ' (focal)'}
-              </TableCell>
-              <TableCell
-                data-testid="insurance.businessArrangement"
-                className="col-md-4"
-              >
-                {insurance.businessArrangement ? (
-                  insurance.businessArrangement
-                ) : (
-                  <MissingValue />
-                )}
-              </TableCell>
-              <TableCell
-                data-testid="insurance.claimResponse"
-                className="col-md-4"
-              >
-                {insurance.claimResponse ? (
-                  <Reference fhirData={insurance.claimResponse} />
-                ) : (
-                  <MissingValue />
-                )}
-              </TableCell>
-            </TableRow>
-          ))}
-        </tbody>
-      </Table>
-    </ValueSection>
-  );
-};
-
-const Item = props => {
-  const {
-    item,
-    parentSequences,
-    collapsedClassName,
-    className,
-    parentId,
-  } = props;
-
-  const itemSequences = [...parentSequences, item.sequence];
-  const id = itemSequences.join('.');
-  const collapseId = parentSequences.length ? parentId : item.sequence;
-
-  const [rotate, setRotate] = useState(false);
-  const handleTableExpand = () => setRotate(!rotate);
-
-  const isExpandable =
-    parentSequences.length === 0 && item.subItems.length !== 0;
-
-  return (
-    <>
-      <TableRow
-        className={`${className} ${collapsedClassName} ${
-          isExpandable ? 'fw-bold table-expandable-row border-top' : ''
-        }`}
-      >
-        <TableCell data-testid="items.sequence" className="col-md-2">
-          {id}
-        </TableCell>
-        <TableCell data-testid="items.service" className="col-md-2">
-          <Coding fhirData={item.service} />
-        </TableCell>
-        <TableCell data-testid="items.unitPrice" className="col-md-2">
-          {item.unitPrice ? (
-            <Money fhirData={item.unitPrice} />
-          ) : (
-            <MissingValue />
-          )}
-          {item.factor != null ? (
-            <span>&nbsp;&times;&nbsp;{item.factor}</span>
-          ) : null}
-        </TableCell>
-        <TableCell data-testid="items.quantity" className="col-md-2">
-          {item.quantity != null ? item.quantity : <MissingValue />}
-        </TableCell>
-        <TableCell data-testid="items.net" className="col-md-2">
-          {item.net ? <Money fhirData={item.net} /> : <MissingValue />}
-        </TableCell>
-        {isExpandable ? (
-          <TableCell className="col-md-2">
-            <button
-              className="fhir-container__Accordion__header-button w-100 p-0 border-0 rounded-1 collapsed text-dark bg-transparent shadow-none point"
-              type="button"
-              data-bs-target={`.item-${collapseId}`}
-              data-bs-toggle={'collapse'}
-              aria-controls={id}
-              aria-expanded="false"
-              onClick={handleTableExpand}
-            >
-              <div className={` ${rotate ? ' header-rotate' : ''}`}>
-                <Chevron strokeColor={rotate ? '#2a6fd7' : '#6f83a9'} />
-              </div>
-            </button>
-          </TableCell>
-        ) : (
-          <TableCell style={{ width: '64px' }} className="col-md-2" />
-        )}
-      </TableRow>
-      {item.subItems.map((subItem, idx) => (
-        <Item
-          key={idx}
-          className="collapse"
-          collapsedClassName={`item-${collapseId}`}
-          item={subItem}
-          parentSequences={itemSequences}
-          parentId={collapseId}
-        />
-      ))}
-    </>
-  );
-};
-
-const Items = props => {
-  const { items } = props;
-
-  return (
-    <ValueSection label="Items" data-testid="items" marginTop>
-      <Table>
-        <thead>
-          <TableRow>
-            <TableHeader>ID</TableHeader>
-            <TableHeader>Service</TableHeader>
-            <TableHeader>Unit price</TableHeader>
-            <TableHeader>Quantity</TableHeader>
-            <TableHeader>Total</TableHeader>
-            <TableHeader></TableHeader>
-          </TableRow>
-        </thead>
-        <tbody className="border-top-0">
-          {items.map((item, idx) => (
-            <Item key={idx} item={item} parentSequences={[]} />
-          ))}
-        </tbody>
-      </Table>
-    </ValueSection>
-  );
-};
-
 const Claim = ({ fhirResource, fhirVersion, fhirIcons }) => {
   let fhirResourceData = {};
   try {
@@ -686,6 +440,8 @@ const Claim = ({ fhirResource, fhirVersion, fhirIcons }) => {
       status: priorityCoding,
     },
   ];
+
+  console.log({ claim: items });
 
   return (
     <Root name="Claim">
